requirements.txt
# Core ML / NLP
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2

# LLM (local, free)
ollama>=0.2.0

# Document processing
spacy>=3.7.0
python-dotenv>=1.0.0

# Streamlit UI
streamlit>=1.35.0

# PDF / document support
pypdf>=4.2.0
python-docx>=1.1.0

# Notebook
ipykernel>=6.29.0
notebook>=7.2.0

# Testing
pytest>=8.2.0
pytest-cov>=5.0.0

# Utilities
numpy>=1.26.0
tqdm>=4.66.0
